#15 Removed node-cron
