+++ app/package.json
@@ -8,11 +8,9 @@
     "dotenv": "^16.0.0",
     "firebase-admin": "^10.1.0",
     "log4js": "^6.4.5",
-    "node-cron": "^3.0.0",
     "redis": "^4.2.0",
     "twitter-api-v2": "^1.12.0"
   },
-  "devDependencies": {},
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
